requirements.txt
blinker==1.7.0
click==8.1.7
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
importlib_metadata==7.1.0
itsdangerous==2.2.0
Jinja2==3.1.3
MarkupSafe==2.1.5
PyJWT==2.8.0
SQLAlchemy==2.0.29
typing_extensions==4.11.0
Werkzeug==3.0.2
zipp==3.18.1
psycopg2==2.9.9
psycopg2-binary==2.9.9